refactor ee proof executor
