+++ crates/proof-impl/evm-ee-stf/src/lib.rs
@@ -1,137 +1,16 @@
 //! EVM Execution Environment STF for Alpen prover, using RSP for EVM execution. Provides primitives
 //! and utilities to process Ethereum block transactions and state transitions in a zkVM.
+pub mod executor;
 pub mod primitives;
 pub mod program;
 pub mod utils;
-use std::{panic, sync::Arc};
 
-use alloy_consensus::{BlockHeader, EthBlock, Header, TxReceipt};
-use alpen_reth_evm::{collect_withdrawal_intents, evm::AlpenEvmFactory};
 pub use primitives::{EvmBlockStfInput, EvmBlockStfOutput};
-use reth_chainspec::ChainSpec;
-use reth_evm::execute::{BasicBlockExecutor, ExecutionOutcome, Executor};
-use reth_evm_ethereum::EthEvmConfig;
-use reth_primitives::EthPrimitives;
-use reth_primitives_traits::block::Block;
-use reth_trie::KeccakKeyHasher;
-use revm::database::WrapDatabaseRef;
-use revm_primitives::alloy_primitives::Bloom;
-use rsp_client_executor::{
-    executor::ClientExecutor, io::EthClientExecutorInput, BlockValidator, FromInput,
-};
+use rsp_client_executor::io::EthClientExecutorInput;
 use utils::generate_exec_update;
 use zkaleido::ZkVmEnv;
 
-pub type AlpEthClientExecutor = ClientExecutor<EthEvmConfig<AlpenEvmFactory>, ChainSpec>;
-
-pub fn process_block_transaction(mut input: EthClientExecutorInput) -> EvmBlockStfOutput {
-    let chain_spec: Arc<ChainSpec> = Arc::new((&input.genesis).try_into().unwrap());
-    let evm_config =
-        EthEvmConfig::new_with_evm_factory(chain_spec.clone(), AlpenEvmFactory::default());
-    // Initialize the witnessed database with verified storage proofs.
-    let db = WrapDatabaseRef(input.witness_db().unwrap());
-
-    let block_executor = BasicBlockExecutor::new(evm_config, db);
-
-    let block = EthPrimitives::from_input_block(input.current_block.clone())
-        .try_into_recovered()
-        .expect("Failed to convert input block");
-
-    // Validate the block header
-    EthPrimitives::validate_header(block.sealed_block().sealed_header(), chain_spec.clone())
-        .expect("Failed to validate block header");
-
-    // Execute the block
-    let execution_output = block_executor
-        .execute(&block)
-        .expect("Failed to execute block");
-
-    // Validate the block post execution.
-    EthPrimitives::validate_block_post_execution(&block, chain_spec.clone(), &execution_output)
-        .expect("Failed to validate block post execution");
-
-    // Accumulate the logs bloom.
-    let mut logs_bloom = Bloom::default();
-    execution_output.result.receipts.iter().for_each(|r| {
-        logs_bloom.accrue_bloom(&r.bloom());
-    });
-
-    // Accumulate withdrawal intents from the executed transactions.
-    let withdrawal_intents = {
-        let transactions = block.into_transactions();
-        let executed_txns = transactions.iter();
-        let receipts_vec = execution_output.receipts.clone();
-        let receipts = receipts_vec.iter();
-        let tx_receipt_pairs = executed_txns.zip(receipts);
-        collect_withdrawal_intents(tx_receipt_pairs).collect::<Vec<_>>()
-    };
-
-    // Convert the output to an execution outcome.
-    let executor_outcome = ExecutionOutcome::new(
-        execution_output.state,
-        vec![execution_output.result.receipts],
-        input.current_block.number,
-        vec![execution_output.result.requests],
-    );
-
-    // Verify the state root.
-    let state_root = {
-        input
-            .parent_state
-            .update(&executor_outcome.hash_state_slow::<KeccakKeyHasher>());
-        input.parent_state.state_root()
-    };
-
-    if state_root != input.current_block.header().state_root() {
-        panic!(
-            "State root mismatch: expected {}, got {}",
-            input.current_block.header().state_root(),
-            state_root
-        );
-    }
-
-    // Derive the block header.
-    // Note: the receipts root and gas used are verified by `validate_block_post_execution`.
-    let header = Header {
-        parent_hash: input.current_block.header().parent_hash(),
-        ommers_hash: input.current_block.header().ommers_hash(),
-        beneficiary: input.current_block.header().beneficiary(),
-        state_root,
-        transactions_root: input.current_block.header().transactions_root(),
-        receipts_root: input.current_block.header().receipts_root(),
-        logs_bloom,
-        difficulty: input.current_block.header().difficulty(),
-        number: input.current_block.header().number(),
-        gas_limit: input.current_block.header().gas_limit(),
-        gas_used: input.current_block.header().gas_used(),
-        timestamp: input.current_block.header().timestamp(),
-        extra_data: input.current_block.header().extra_data().clone(),
-        mix_hash: input.current_block.header().mix_hash().unwrap(),
-        nonce: input.current_block.header().nonce().unwrap(),
-        base_fee_per_gas: input.current_block.header().base_fee_per_gas(),
-        withdrawals_root: input.current_block.header().withdrawals_root(),
-        blob_gas_used: input.current_block.header().blob_gas_used(),
-        excess_blob_gas: input.current_block.header().excess_blob_gas(),
-        parent_beacon_block_root: input.current_block.header().parent_beacon_block_root(),
-        requests_hash: input.current_block.header().requests_hash(),
-    };
-
-    let deposit_requests = input
-        .current_block
-        .withdrawals()
-        .map(|w| w.to_vec())
-        .unwrap_or_default();
-
-    EvmBlockStfOutput {
-        block_idx: header.number,
-        new_blockhash: header.hash_slow(),
-        new_state_root: header.state_root,
-        prev_blockhash: header.parent_hash,
-        txn_root: header.transactions_root,
-        deposit_requests,
-        withdrawal_intents,
-    }
-}
+use crate::executor::process_block;
 
 /// Processes a sequence of EL block transactions from the given `zkvm` environment, ensuring block
 /// hash continuity and committing the resulting updates.
@@ -144,7 +23,7 @@ pub fn process_block_transaction_outer(zkvm: &impl ZkVmEnv) {
 
     for _ in 0..num_blocks {
         let input: EthClientExecutorInput = zkvm.read_serde();
-        let output = process_block_transaction(input);
+        let output = process_block(input).expect("Failed to process block transaction");
 
         if let Some(expected_hash) = current_blockhash {
             assert_eq!(output.prev_blockhash, expected_hash, "Block hash mismatch");
@@ -162,7 +41,7 @@ mod tests {
 
     use serde::{Deserialize, Serialize};
 
-    use super::{process_block_transaction, EvmBlockStfInput, EvmBlockStfOutput};
+    use super::{process_block, EvmBlockStfInput, EvmBlockStfOutput};
 
     #[derive(Serialize, Deserialize)]
     struct TestData {
@@ -195,7 +74,7 @@ mod tests {
         let test_data = get_mock_data();
 
         let input = test_data.witness;
-        let op = process_block_transaction(input);
+        let op = process_block(input).expect("Failed to process block transaction");
         assert_eq!(op, test_data.params);
     }
 }
